redone users and localauths migration to maintain versioning
